Fix TaskControllerTests
