--- AllReadyApp/Web-App/AllReady.UnitTest/Controllers/TaskControllerTests.cs
@@ -1,943 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using AllReady.Areas.Admin.Controllers;
-using AllReady.Models;
-using Microsoft.AspNetCore.Mvc;
-using Moq;
-using Xunit;
-using Microsoft.AspNetCore.Http;
-using System.Security.Claims;
-using AllReady.UnitTest.Extensions;
-using MediatR;
-using Microsoft.AspNetCore.Authorization;
-using AllReady.Areas.Admin.Features.Tasks;
-using System.Threading.Tasks;
-using AllReady.Features.Event;
-using AllReady.Extensions;
-using AllReady.Areas.Admin.ViewModels.Task;
-using AllReady.Areas.Admin.ViewModels.Validators;
-
-namespace AllReady.UnitTest.Controllers
-{
-    public class TaskControllerTests
-    {
-        [Fact]
-        public void CreateGetSendsEventByEventIdQueryWithTheCorrectEventId()
-        {
-            const int eventId = 1;
-            var mediator = new Mock<IMediator>();
-            var mockValidator = new Mock<ITaskEditViewModelValidator>();
-            var sut = new TaskController(mediator.Object, mockValidator.Object);
-            sut.Create(eventId);
-
-            mediator.Verify(x => x.Send(It.Is<EventByIdQuery>(y => y.EventId == eventId)), Times.Once);
-        }
-
-        [Fact]
-        public void CreateGetReturnsHttpUnauthorizedResultWhenEventIsNull()
-        {
-            var sut = new TaskController(Mock.Of<IMediator>(), Mock.Of<ITaskEditViewModelValidator>());
-            var result = sut.Create(It.IsAny<int>());
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public void CreateGetReturnsHttpUnauthorizedResultWhenUserIsNotOrganizationAdminForTheirOrganization()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { ManagingOrganizationId = 1 } });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-
-            var result = sut.Create(It.IsAny<int>());
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public void CreateGetReturnsCorrectViewModel()
-        {
-            const int organizationId = 1;
-            const int campaignId = 1;
-            var campaignStartDateTime = DateTime.Now.AddDays(-1);
-            var campaignEndDateTime = DateTime.Now.AddDays(1);
-
-            var campaignEvent = new Event
-            {
-                Id = 1,
-                Name = "EventName",
-                CampaignId = campaignId,
-                Campaign = new Campaign
-                {
-                    Id = campaignId,
-                    Name = "CampaignName",
-                    ManagingOrganizationId = organizationId,
-                    TimeZoneId = "EST",
-                    StartDateTime = campaignStartDateTime,
-                    EndDateTime = campaignEndDateTime
-                }
-            };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(campaignEvent);
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = sut.Create(It.IsAny<int>()) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            Assert.Equal(modelResult.EventId, campaignEvent.Id);
-            Assert.Equal(modelResult.EventName, campaignEvent.Name);
-            Assert.Equal(modelResult.CampaignId, campaignEvent.CampaignId);
-            Assert.Equal(modelResult.CampaignName, campaignEvent.Campaign.Name);
-            Assert.Equal(modelResult.OrganizationId, campaignEvent.Campaign.ManagingOrganizationId);
-            Assert.Equal(modelResult.TimeZoneId, campaignEvent.Campaign.TimeZoneId);
-            Assert.Equal(modelResult.StartDateTime, campaignEvent.StartDateTime);
-            Assert.Equal(modelResult.EndDateTime, campaignEvent.EndDateTime);
-        }
-
-        [Fact]
-        public void CreateGetReturnsCorrectView()
-        {
-            const int organizationId = 1;
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { ManagingOrganizationId = organizationId } });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = sut.Create(It.IsAny<int>()) as ViewResult;
-
-            Assert.Equal(result.ViewName, "Edit");
-        }
-
-        [Fact]
-        public void CreateGetHasHttpGetAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Create(It.IsAny<int>())).OfType<HttpGetAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void CreateGetHasRouteAttributeWithCorrectTemplate()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Create(It.IsAny<int>())).OfType<RouteAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.Template, "Admin/Task/Create/{eventId}");
-        }
-
-        [Fact]
-        public async Task CreatePostCallsValidateOnTaskSummaryModelValidator()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" } });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            sut.SetFakeUserType(UserType.OrgAdmin);
-
-            await sut.Create(It.IsAny<int>(), new TaskSummaryViewModel { OrganizationId = 1 });
-
-            validator.Verify(x => x.Validate(It.IsAny<TaskSummaryViewModel>()), Times.Once);
-        }
-
-        [Fact]
-        public async Task CreatePostAddsCorrectModelStateErrorWhenValidatorReturnsError()
-        {
-            var mediator = new Mock<IMediator>();
-            //mgmccarthy: for a list of time zone's to feed to TimeZoneInfo.FindSystemTimeZoneById, check here:
-            //http://stackoverflow.com/questions/7908343/list-of-timezone-ids-for-use-with-findtimezonebyid-in-c
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" } });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>> { new KeyValuePair<string, string>(nameof(TaskSummaryViewModel.EndDateTime), "Ending time cannot be earlier than the starting time") });
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            await sut.Create(It.IsAny<int>(), new TaskSummaryViewModel { EndDateTime = DateTimeOffset.Now.AddDays(-1), StartDateTime = DateTimeOffset.Now.AddDays(1) });
-
-            var modelStateErrorCollection = sut.ModelState.GetErrorMessagesByKey(nameof(TaskSummaryViewModel.EndDateTime));
-            Assert.Equal(modelStateErrorCollection.Single().ErrorMessage, "Ending time cannot be earlier than the starting time");
-        }
-
-        [Fact]
-        public async Task CreatePostReturnsCorrectViewAndViewModelWhenValidatorReturnsError()
-        {
-            var model = new TaskSummaryViewModel { EndDateTime = DateTimeOffset.Now.AddDays(-1), StartDateTime = DateTimeOffset.Now.AddDays(1) };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" } });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>> { new KeyValuePair<string, string>(nameof(TaskSummaryViewModel.EndDateTime), "Ending time cannot be earlier than the starting time") });
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            var result = await sut.Create(It.IsAny<int>(), model) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            Assert.IsType<ViewResult>(result);
-            Assert.Equal(modelResult, model);
-        }
-
-        [Fact]
-        public async Task CreatePostReturnsHttpUnauthorizedResultWhenModelStateIsValidAndUserIsNotAnOrganizationAdminUser()
-        {
-            var startDateTime = DateTimeOffset.Now.AddDays(-1);
-            var endDateTime = DateTimeOffset.Now.AddDays(1);
-            var model = new TaskSummaryViewModel { StartDateTime = startDateTime, EndDateTime = endDateTime, OrganizationId = 1 };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event
-            {
-                Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" },
-                StartDateTime = startDateTime.AddDays(-1),
-                EndDateTime = endDateTime.AddDays(1)
-            });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            sut.SetDefaultHttpContext();
-
-            var result = await sut.Create(It.IsAny<int>(), model);
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task CreatePostSendsEditTaskCommandAsyncWhenModelStateIsValidAndUserIsOrganizationAdmin()
-        {
-            const int organizationId = 1;
-            var taskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId };
-            var startDateTime = DateTimeOffset.Now.AddDays(-1);
-            var endDateTime = DateTimeOffset.Now.AddDays(1);
-
-            var model = new TaskSummaryViewModel { StartDateTime = startDateTime, EndDateTime = endDateTime, OrganizationId = organizationId };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event
-            {
-                Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" },
-                StartDateTime = startDateTime.AddDays(-1),
-                EndDateTime = endDateTime.AddDays(1)
-            });
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            await sut.Create(It.IsAny<int>(), model);
-
-            mediator.Verify(x => x.SendAsync(It.Is<EditTaskCommandAsync>(y => y.Task == model)));
-        }
-
-        [Fact]
-        public async Task CreatePostRedirectsToCorrectAction()
-        {
-            const int organizationId = 1;
-            var taskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId };
-            var startDateTime = DateTimeOffset.Now.AddDays(-1);
-            var endDateTime = DateTimeOffset.Now.AddDays(1);
-            const int TaskSummaryModelId = 1;
-            var model = new TaskSummaryViewModel { Id = TaskSummaryModelId, StartDateTime = startDateTime, EndDateTime = endDateTime, OrganizationId = organizationId };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event
-            {
-                Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" },
-                StartDateTime = startDateTime.AddDays(-1),
-                EndDateTime = endDateTime.AddDays(1)
-            });
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var routeValues = new Dictionary<string, object> { ["id"] = model.Id };
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = await sut.Create(TaskSummaryModelId, model) as RedirectToActionResult;
-
-            Assert.Equal(result.ActionName, nameof(TaskController.Details));
-            Assert.Equal(result.ControllerName, "Event");
-            Assert.Equal(result.RouteValues, routeValues);
-        }
-
-        [Fact]
-        public void CreatePostHasHttpGetAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Create(It.IsAny<int>(), It.IsAny<TaskSummaryViewModel>())).OfType<HttpPostAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void CreatePostHasRouteAttributeWithCorrectTemplate()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Create(It.IsAny<int>(), It.IsAny<TaskSummaryViewModel>())).OfType<RouteAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.Template, "Admin/Task/Create/{eventId}");
-        }
-
-        [Fact]
-        public async Task EditGetSendsEditTaskQueryAsyncWithCorrectTaskId()
-        {
-            const int taskId = 1;
-            var mediator = new Mock<IMediator>();
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            await sut.Edit(taskId);
-
-            mediator.Verify(x => x.SendAsync(It.Is<EditTaskQueryAsync>(y => y.TaskId == taskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task EditGetReturnsHttpNotFoundResultWhenTaskIsNull()
-        {
-            var sut = new TaskController(Mock.Of<IMediator>(), Mock.Of<ITaskEditViewModelValidator>());
-            var result = await sut.Edit(It.IsAny<int>());
-
-            Assert.IsType<NotFoundResult>(result);
-        }
-
-        [Fact]
-        public async Task EditGetReturnsHttpUnauthorizedResultWhenUserIsNotOrgAdmin()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<EditTaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel());
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-            var result = await sut.Edit(It.IsAny<int>());
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task EditGetReturnsCorrectViewModelAndView()
-        {
-            const int organizationId = 1;
-            var TaskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<EditTaskQueryAsync>())).ReturnsAsync(TaskSummaryModel);
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = await sut.Edit(It.IsAny<int>()) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            Assert.IsType<ViewResult>(result);
-            Assert.IsType<TaskSummaryViewModel>(modelResult);
-            Assert.Equal(modelResult, TaskSummaryModel);
-        }
-
-        [Fact]
-        public void EditGetHasHttpGetAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Edit(It.IsAny<int>())).OfType<HttpGetAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void EditGetHasRouteAttributeWithCorrectTemplate()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Edit(It.IsAny<int>())).OfType<RouteAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.Template, "Admin/Task/Edit/{id}");
-        }
-
-        [Fact]
-        public async Task EditPostCallsValidateOnTaskSummaryModelValidator()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" } });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            sut.SetFakeUserType(UserType.OrgAdmin);
-
-            await sut.Edit(new TaskSummaryViewModel { EndDateTime = DateTimeOffset.Now.AddDays(-1), StartDateTime = DateTimeOffset.Now.AddDays(1), OrganizationId = 1 });
-
-            validator.Verify(x => x.Validate(It.IsAny<TaskSummaryViewModel>()), Times.Once);
-        }
-
-        [Fact]
-        public async Task EditPostAddsCorrectModelStateErrorWhenValidatorReturnsError()
-        {
-            var model = new TaskSummaryViewModel { EndDateTime = DateTimeOffset.Now.AddDays(-1), StartDateTime = DateTimeOffset.Now.AddDays(1) };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" } });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>> { new KeyValuePair<string, string>(nameof(TaskSummaryViewModel.EndDateTime), "Ending time cannot be earlier than the starting time") });
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            var result = await sut.Edit(model) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            var modelStateErrorCollection = sut.ModelState.GetErrorMessagesByKey(nameof(TaskSummaryViewModel.EndDateTime));
-            Assert.Equal(modelStateErrorCollection.Single().ErrorMessage, "Ending time cannot be earlier than the starting time");
-        }
-
-        [Fact]
-        public async Task EditPostReturnsCorrectViewAndViewModelWhenValidatorReturnsError()
-        {
-            var model = new TaskSummaryViewModel { EndDateTime = DateTimeOffset.Now.AddDays(-1), StartDateTime = DateTimeOffset.Now.AddDays(1) };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" } });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>> { new KeyValuePair<string, string>(nameof(TaskSummaryViewModel.EndDateTime), "Ending time cannot be earlier than the starting time") });
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            var result = await sut.Edit(model) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            Assert.IsType<ViewResult>(result);
-            Assert.Equal(modelResult, model);
-        }
-
-        [Fact]
-        public async Task EditPostReturnsHttpUnauthorizedResultWhenModelStateIsValidAndUserIsNotAnOrganizationAdminUser()
-        {
-            var startDateTime = DateTimeOffset.Now.AddDays(-1);
-            var endDateTime = DateTimeOffset.Now.AddDays(1);
-            var model = new TaskSummaryViewModel { StartDateTime = startDateTime, EndDateTime = endDateTime, OrganizationId = 1 };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event
-            {
-                Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" },
-                StartDateTime = startDateTime.AddDays(-1),
-                EndDateTime = endDateTime.AddDays(1)
-            });
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            sut.SetDefaultHttpContext();
-
-            var result = await sut.Edit(model);
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task EditPostSendsEditTaskCommandAsyncWhenModelStateIsValidAndUserIsOrganizationAdmin()
-        {
-            const int organizationId = 1;
-            var taskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId };
-            var startDateTime = DateTimeOffset.Now.AddDays(-1);
-            var endDateTime = DateTimeOffset.Now.AddDays(1);
-
-            var model = new TaskSummaryViewModel { StartDateTime = startDateTime, EndDateTime = endDateTime, OrganizationId = organizationId };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event
-            {
-                Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" },
-                StartDateTime = startDateTime.AddDays(-1),
-                EndDateTime = endDateTime.AddDays(1)
-            });
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            await sut.Edit(model);
-
-            mediator.Verify(x => x.SendAsync(It.Is<EditTaskCommandAsync>(y => y.Task == model)));
-        }
-
-        [Fact]
-        public async Task EditPostRedirectsToCorrectAction()
-        {
-            const int organizationId = 1;
-            var taskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId };
-            var startDateTime = DateTimeOffset.Now.AddDays(-1);
-            var endDateTime = DateTimeOffset.Now.AddDays(1);
-            var model = new TaskSummaryViewModel { Id = 1, StartDateTime = startDateTime, EndDateTime = endDateTime, OrganizationId = organizationId };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event
-            {
-                Campaign = new Campaign { TimeZoneId = "Eastern Standard Time" },
-                StartDateTime = startDateTime.AddDays(-1),
-                EndDateTime = endDateTime.AddDays(1)
-            });
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var validator = new Mock<ITaskEditViewModelValidator>();
-            validator.Setup(x => x.Validate(It.IsAny<TaskSummaryViewModel>())).Returns(new List<KeyValuePair<string, string>>());
-
-            var sut = new TaskController(mediator.Object, validator.Object);
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = await sut.Edit(model) as RedirectToActionResult;
-
-            var routeValues = new Dictionary<string, object> { ["id"] = model.Id };
-
-            Assert.Equal(result.ControllerName, "Task");
-            Assert.Equal(result.ActionName, nameof(TaskController.Details));
-            Assert.Equal(result.RouteValues, routeValues);
-        }
-
-        [Fact]
-        public void EditPostHasHttpPostAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Edit(It.IsAny<TaskSummaryViewModel>())).OfType<HttpPostAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void EditPostHasValidateAntiForgeryTokenAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Edit(It.IsAny<TaskSummaryViewModel>())).OfType<ValidateAntiForgeryTokenAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public async Task DeleteSendsTaskQueryAsyncWithCorrectTaskId()
-        {
-            const int taskId = 1;
-            var mediator = new Mock<IMediator>();
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            await sut.Delete(taskId);
-
-            mediator.Verify(x => x.SendAsync(It.Is<TaskQueryAsync>(y => y.TaskId == taskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task DeleteReturnsHttpNotFoundResultWhenTaskIsNull()
-        {
-            var sut = new TaskController(Mock.Of<IMediator>(), Mock.Of<ITaskEditViewModelValidator>());
-            var result = await sut.Delete(It.IsAny<int>());
-            Assert.IsType<NotFoundResult>(result);
-        }
-
-        [Fact]
-        public async Task DeleteReturnsHttpUnauthorizedResultWhenUserIsNotOrgAdmin()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel());
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-            var result = await sut.Delete(It.IsAny<int>());
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task DeleteReturnsCorrectViewModelAndView()
-        {
-            const int organizationId = 1;
-            var taskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = await sut.Delete(It.IsAny<int>()) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            Assert.IsType<ViewResult>(result);
-            Assert.IsType<TaskSummaryViewModel>(modelResult);
-            Assert.Equal(modelResult, taskSummaryModel);
-        }
-
-        [Fact]
-        public async Task DetailsSendsTaskQueryAsyncWithCorrectTaskId()
-        {
-            const int taskId = 1;
-            var mediator = new Mock<IMediator>();
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            await sut.Details(taskId);
-
-            mediator.Verify(x => x.SendAsync(It.Is<TaskQueryAsync>(y => y.TaskId == taskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task DetailsReturnsHttpNotFoundResultWhenTaskIsNull()
-        {
-            var sut = new TaskController(Mock.Of<IMediator>(), Mock.Of<ITaskEditViewModelValidator>());
-            var result = await sut.Details(It.IsAny<int>());
-            Assert.IsType<NotFoundResult>(result);
-        }
-
-        [Fact]
-        public async Task DetailsReturnsCorrectViewModelAndView()
-        {
-            var taskSummaryModel = new TaskSummaryViewModel();
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            var result = await sut.Details(It.IsAny<int>()) as ViewResult;
-            var modelResult = result.ViewData.Model as TaskSummaryViewModel;
-
-            Assert.IsType<ViewResult>(result);
-            Assert.IsType<TaskSummaryViewModel>(modelResult);
-            Assert.Equal(modelResult, taskSummaryModel);
-        }
-
-        [Fact]
-        public void DetailsHasHttpGetAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Details(It.IsAny<int>())).OfType<HttpGetAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void DetailsHasRouteAttributeWithCorrectTemplate()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Details(It.IsAny<int>())).OfType<RouteAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.Template, "Admin/Task/Details/{id}");
-        }
-
-        [Fact]
-        public async Task DeleteConfirmedSendsTaskQueryAsyncWithCorrectTaskId()
-        {
-            const int taskId = 1;
-            var mediator = new Mock<IMediator>();
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            await sut.DeleteConfirmed(taskId);
-
-            mediator.Verify(x => x.SendAsync(It.Is<TaskQueryAsync>(y => y.TaskId == taskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task DeleteConfirmedReturnsHttpNotFoundResultWhenTaskSummaryModelIsNull()
-        {
-            var sut = new TaskController(Mock.Of<IMediator>(), Mock.Of<ITaskEditViewModelValidator>());
-            var result = await sut.DeleteConfirmed(It.IsAny<int>());
-            Assert.IsType<NotFoundResult>(result);
-        }
-
-        [Fact]
-        public async Task DeleteConfirmedReturnsHttpUnauthorizedResultWhenUserIsNotOrgAdmin()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel());
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-
-            var result = await sut.DeleteConfirmed(It.IsAny<int>());
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task DeleteConfirmedSendsDeleteTaskCommandAsyncWithCorrectTaskId()
-        {
-            const int organizationId = 1;
-            const int taskId = 1;
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel { OrganizationId = 1 });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-            await sut.DeleteConfirmed(taskId);
-
-            mediator.Verify(x => x.SendAsync(It.Is<DeleteTaskCommandAsync>(y => y.TaskId == taskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task DeleteConfirmedRedirectsToCorrectAction()
-        {
-            const int organizationId = 1;
-            const int taskId = 1;
-            var taskSummaryModel = new TaskSummaryViewModel { OrganizationId = organizationId, EventId = 1 };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskSummaryModel);
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-            var result = await sut.DeleteConfirmed(taskId) as RedirectToActionResult;
-
-            var routeValues = new Dictionary<string, object> { ["id"] = taskSummaryModel.EventId };
-
-            Assert.Equal(result.ActionName, nameof(EventController.Details));
-            Assert.Equal(result.ControllerName, "Event");
-            Assert.Equal(result.RouteValues, routeValues);
-        }
-
-        [Fact]
-        public void DeleteConfirmedHasHttpPostAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.DeleteConfirmed(It.IsAny<int>())).OfType<HttpPostAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        public void DeleteConfirmedHasActionNameAttributeWithCorrectActionName()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.DeleteConfirmed(It.IsAny<int>())).OfType<ActionNameAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.Name, "Delete");
-        }
-
-        [Fact]
-        public void DeleteConfirmedHasValidateAntiForgeryTokenAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.DeleteConfirmed(It.IsAny<int>())).OfType<ValidateAntiForgeryTokenAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public async Task AssignSendsTaskQueryAsyncWithCorrectTaskId()
-        {
-            const int taskId = 1;
-            var taskModelSummary = new TaskSummaryViewModel { EventId = 1 };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { ManagingOrganizationId = 1 } });
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskModelSummary);
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-            await sut.Assign(taskId, null);
-
-            mediator.Verify(x => x.SendAsync(It.Is<TaskQueryAsync>(y => y.TaskId == taskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task AssignReturnsHttpUnauthorizedResultWhenUserIsNotOrgAdmin()
-        {
-            var taskModelSummary = new TaskSummaryViewModel { EventId = 1 };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskModelSummary);
-            mediator.Setup(x => x.Send(It.IsAny<EventByIdQuery>())).Returns(new Event { Campaign = new Campaign { ManagingOrganizationId = 1 } });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-            var result = await sut.Assign(1, null);
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task AssignSendsAssignTaskCommandAsync()
-        {
-            const int organizationId = 1;
-            const int taskId = 1;
-            var taskModelSummary = new TaskSummaryViewModel { EventId = 1 };
-            var userIds = new List<string> { "1", "2" };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskModelSummary);
-            mediator.Setup(x => x.Send(It.Is<EventByIdQuery>(y => y.EventId == taskModelSummary.EventId))).Returns(new Event { Campaign = new Campaign { ManagingOrganizationId = organizationId } });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-            await sut.Assign(taskId, userIds);
-
-            mediator.Verify(x => x.SendAsync(It.Is<AssignTaskCommandAsync>(y => y.TaskId == taskId && y.UserIds == userIds)), Times.Once);
-        }
-
-        [Fact]
-        public async Task AssignRedirectsToRoute()
-        {
-            const int organizationId = 1;
-            const int taskId = 1;
-            var taskModelSummary = new TaskSummaryViewModel { EventId = 1 };
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(taskModelSummary);
-            mediator.Setup(x => x.Send(It.Is<EventByIdQuery>(y => y.EventId == taskModelSummary.EventId))).Returns(new Event { Campaign = new Campaign { ManagingOrganizationId = organizationId } });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-
-            var result = await sut.Assign(taskId, null) as RedirectToRouteResult;
-
-            Assert.Equal(result.RouteValues["controller"], "Task");
-            Assert.Equal(result.RouteValues["Area"], "Admin");
-            Assert.Equal(result.RouteValues["action"], nameof(TaskController.Details));
-            Assert.Equal(result.RouteValues["id"], taskId);
-        }
-
-        [Fact]
-        public void AssignHasHttpPostAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Assign(It.IsAny<int>(), It.IsAny<List<string>>())).OfType<HttpPostAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void AssignHasValidateAntiForgeryTokenAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.Assign(It.IsAny<int>(), It.IsAny<List<string>>())).OfType<ValidateAntiForgeryTokenAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public async Task MessageAllVolunteersReturnsBadRequestObjectResultWhenModelStateIsInvalid()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            sut.AddModelStateError();
-            var result = await sut.MessageAllVolunteers(It.IsAny<MessageTaskVolunteersViewModel>());
-            Assert.IsType<BadRequestObjectResult>(result);
-        }
-
-        [Fact]
-        public async Task MessageAllVounteersSendsTaskQueryAsyncWithCorrectTaskId()
-        {
-            var model = new MessageTaskVolunteersViewModel { TaskId = 1 };
-            var mediator = new Mock<IMediator>();
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            await sut.MessageAllVolunteers(model);
-
-            mediator.Verify(x => x.SendAsync(It.Is<TaskQueryAsync>(y => y.TaskId == model.TaskId)), Times.Once);
-        }
-
-        [Fact]
-        public async Task MessageAllVolunteersReturnsHttpNotFoundResultWhenTaskIsNull()
-        {
-            var mediator = new Mock<IMediator>();
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            var result = await sut.MessageAllVolunteers(new MessageTaskVolunteersViewModel());
-
-            Assert.IsType<NotFoundResult>(result);
-        }
-
-        [Fact]
-        public async Task MessageAllVolunttersReturnsHttpUnauthorizedResultWhenUserIsNotOrgAdmin()
-        {
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel());
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-            var result = await sut.MessageAllVolunteers(new MessageTaskVolunteersViewModel());
-
-            Assert.IsType<UnauthorizedResult>(result);
-        }
-
-        [Fact]
-        public async Task MessageAllVolunteersSendsMessageTaskVolunteersCommandAsyncWithCorrectData()
-        {
-            const int organizationId = 1;
-            var model = new MessageTaskVolunteersViewModel();
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel { OrganizationId = organizationId });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-            await sut.MessageAllVolunteers(model);
-
-            mediator.Verify(x => x.SendAsync(It.Is<MessageTaskVolunteersCommandAsync>(y => y.Model == model)), Times.Once);
-        }
-
-        [Fact]
-        public async Task MessageAllVolunteersReturnsHttpOkResult()
-        {
-            const int organizationId = 1;
-
-            var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<TaskQueryAsync>())).ReturnsAsync(new TaskSummaryViewModel { OrganizationId = organizationId });
-
-            var sut = new TaskController(mediator.Object, Mock.Of<ITaskEditViewModelValidator>());
-            MakeUserOrganizationAdminUser(sut, organizationId.ToString());
-            var result = await sut.MessageAllVolunteers(new MessageTaskVolunteersViewModel());
-
-            Assert.IsType<OkResult>(result);
-        }
-
-        [Fact]
-        public void MessageAllVolunteersHasHttpPostAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.MessageAllVolunteers(It.IsAny<MessageTaskVolunteersViewModel>())).OfType<HttpPostAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void MessageAllVolunteersHasValidateAntiForgeryTokenAttribute()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributesOn(x => x.MessageAllVolunteers(It.IsAny<MessageTaskVolunteersViewModel>())).OfType<ValidateAntiForgeryTokenAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-        }
-
-        [Fact]
-        public void ControllerHasAreaAtttributeWithTheCorrectAreaName()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            sut.SetDefaultHttpContext();
-            var attribute = sut.GetAttributes().OfType<AreaAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.RouteValue, "Admin");
-        }
-
-        [Fact]
-        public void ControllerHasAuthorizeAtttributeWithTheCorrectPolicy()
-        {
-            var sut = new TaskController(null, Mock.Of<ITaskEditViewModelValidator>());
-            var attribute = sut.GetAttributes().OfType<AuthorizeAttribute>().SingleOrDefault();
-            Assert.NotNull(attribute);
-            Assert.Equal(attribute.Policy, "OrgAdmin");
-        }
-
-        private static void MakeUserOrganizationAdminUser(Controller controller, string organizationId)
-        {
-            var orgAdminClaims = new List<Claim>
-            {
-                new Claim(AllReady.Security.ClaimTypes.UserType, Enum.GetName(typeof(UserType), UserType.OrgAdmin)),
-                new Claim(AllReady.Security.ClaimTypes.Organization, organizationId)
-            };
-
-            var httpContext = new Mock<HttpContext>();
-            var claimsPrincipal = new ClaimsPrincipal(new ClaimsIdentity(orgAdminClaims));
-            httpContext.Setup(x => x.User).Returns(claimsPrincipal);
-
-            controller.ControllerContext.HttpContext = httpContext.Object;
-        }
-    }
-}
