+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Controllers/TaskController.cs
@@ -4,8 +4,6 @@ using System.Threading.Tasks;
 using AllReady.Areas.Admin.Features.Tasks;
 using AllReady.Areas.Admin.ViewModels.Task;
 using AllReady.Areas.Admin.ViewModels.Validators;
-using AllReady.Features.Event;
-using AllReady.Models;
 using AllReady.Security;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
